REVERT: Remove HP system (caused 15-27 FPS performance issues)
The HP system with damage numbers, HP bars, and multi-hit combat
caused severe performance degradation (15-27 FPS instead of 60 FPS).

Reverted to:
- One-shot kills (instant enemy destruction)
- No HP tracking
- No HP bars
- No damage numbers
- Simple particle effects on hit

Performance should now be back to 60 FPS.

The HP system will need to be redesigned with object pooling
and more efficient rendering if we want to add it back in the future.